fix(download): parse new menu date format
Prior to this commit, the format of the Dinner Daily menu dates were
assumed to be `<month> <day>` (e.g., `Oct 28`). However, as of this
commit, the Dinner Daily website has updated to a new date format for
menus, `<month> <start> to <end>` (e.g., `Oct 28 to 03`).

In this commit, fix the `download()` to handle the new date format.