feat(download): simulate user/passwd login
Prior to this commit, menu and shopping list downloads used user browser
cookies for authentication. This sometimes failed due to, e.g., expired
cookies, where it was necessary to manually log in to the Dinner Daily
website again to refresh the cookie.

In this commit, modify download code to authenticate directly with the
Dinner Daily website using a username and password. This ensures that
a new cookie is obtained each time. It also eliminates the need for
access to the user's browser cookies.

Username and password may be set using the following environment
variables, respectively:

 - `DINNER_DAILY_USERNAME`
 - `DINNER_DAILY_PASSWORD`

BREAKING CHANGE: deprecate use of user browser cookies for downloads;
instead require username/password for authentication.