build(appveyor): build current week (not previous)
The original intent behind downloading and building the _previous_ week
was to prevent downloading a menu before one had a chance to customize
it.

In this commit, download and build _current_ week instead. This will
ensure current week's menu is built, while allowing one to manually
trigger a build on Appveyor to overwrite the current week's build
release based on an updated menu.

BREAKING CHANGE: download and build current week (not previous week).